Refactor NPU/DPU setup for future re-use
